feat: 代码重构

+++ src/main/java/com/axr/stockmanage/controller/StockController.java
@@ -2,8 +2,8 @@ package com.axr.stockmanage.controller;
 
 import com.axr.stockmanage.common.BusinessException;
 import com.axr.stockmanage.common.Result;
-import com.axr.stockmanage.model.entity.Stock;
 import com.axr.stockmanage.model.dto.StockDTO;
+import com.axr.stockmanage.model.entity.Stock;
 import com.axr.stockmanage.service.StockService;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,14 +22,14 @@ public class StockController {
 
     @PostMapping("update")
     public Result<Integer> updateStock(@RequestBody StockDTO dto) {
-        if (dto.getProductId() == null || dto.getProductId() < 1) {
+        if (dto.getProductId() == null || dto.getProductId() < 1L) {
             throw new BusinessException("信息错误");
         }
         return Result.success(stockService.updateStock(dto));
     }
 
     @GetMapping("search")
-    public Result<Stock> searchStock(Integer productId) {
+    public Result<Stock> searchStock(Long productId) {
         return Result.success(stockService.getStockByProductId(productId));
     }
 }
